+++ lib/src/sparse_list.dart
@@ -52,8 +52,7 @@ class SparseList<E> extends Object with ListMixin<E> {
   /**
    * Returns a read-only list of the groups.
    */
-  List<GroupedRangeList<E>> get groups =>
-      new UnmodifiableListView<GroupedRangeList<E>>(_groups);
+  List<GroupedRangeList<E>> get groups => new UnmodifiableListView<GroupedRangeList<E>>(_groups);
 
   /**
    * Returns the number of groups.
@@ -205,6 +204,36 @@ class SparseList<E> extends Object with ListMixin<E> {
     if (_length < group.end + 1) {
       _length = group.end + 1;
     }
+  }
+
+  /**
+   * Returns all groups that intersects with the specified [range] which being
+   * expanded (with default value) or shrunk to specified [range].
+   */
+  List<GroupedRangeList<E>> getAlignedGroups(RangeList range) {
+    var groups = getGroups(range).toList();
+    var length = groups.length;
+    if (length == 0) {
+      return [new GroupedRangeList<E>(range.start, range.end, defaultValue)];
+    }
+
+    var first = groups.first;
+    if (range.start > first.start) {
+      groups[0] = first.intersection(range);
+    } else if (range.start < first.start) {
+      var insertion = new GroupedRangeList<E>(range.start, first.start - 1, defaultValue);
+      groups.insert(0, insertion);
+    }
+
+    var last = groups.last;
+    if (range.end > last.end) {
+      var addition = new GroupedRangeList<E>(last.end + 1, range.end, defaultValue);
+      groups.add(addition);
+    } else if (range.end < last.end) {
+      groups[groups.length - 1] = last.intersection(range);
+    }
+
+    return groups;
   }
 
   /**
@@ -474,16 +503,13 @@ class SparseList<E> extends Object with ListMixin<E> {
             return;
           } else {
             if (groupStart <= currentStart) {
-              _groups.insert(
-                  i,
-                  new GroupedRangeList(currentStart, groupEnd, groupKey));
+              _groups.insert(i, new GroupedRangeList(currentStart, groupEnd, groupKey));
               _groups[i + 1] = parts.first;
               affected.add(i);
               break;
             } else {
               _groups.insert(i, parts.first);
-              _groups[i +
-                  1] = new GroupedRangeList(groupStart, currentEnd, groupKey);
+              _groups[i + 1] = new GroupedRangeList(groupStart, currentEnd, groupKey);
               length++;
             }
           }
@@ -528,8 +554,7 @@ class SparseList<E> extends Object with ListMixin<E> {
   }
 }
 
-class _SparseListIndexesIterator extends Object with IterableMixin<int>
-    implements Iterator<int> {
+class _SparseListIndexesIterator extends Object with IterableMixin<int> implements Iterator<int> {
   int _count;
 
   int _current;